+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Release 5.8.0:
    - In `CredentialIssuer` deprecate constructor parameter `credentialProvider`, replace with `credentialDataProvider`
    - Extend `CredentialToBeIssued` to contain properties `expiration`, `scheme`, `subjectPublicKey`, `userInfo`
    - In `CredentialIssuer` move constructor parameters for loading data and issuing credentials to method `credential()`
+   - Extract `ProofValidator` out of `CredentialIssuer`
 
 Release 5.7.1:
  - Signum 3.16.3/Supreme 0.8.3 to fix certificate encoding in JWS header